Place photos

+++ app/controllers/activities_controller.rb
@@ -1,15 +1,27 @@
 class ActivitiesController < ApplicationController
+  before_action :user, only: [:create]
+
   def index
-    @activites = Activity.all
+    @activities = Activity.all
   end
 
   def show
     @activity = Activity.find(params[:id])
+    authorize @activity
+  end
+
+  def new
+    @activity = Activity.new
+    authorize @activity
   end
 
   def create
     @activity = Activity.new(activity_params)
-    @activity.user = @user
+    @activity.owner = @user
+    @place = Place.find(params[:activity][:place])
+    @activity.place = @place
+    authorize @activity
+
     if @activity.save
       redirect_to activity_path(@activity)
     else
@@ -17,27 +29,27 @@ class ActivitiesController < ApplicationController
     end
   end
 
-  def new
-    @activity = Activity.new
-  end
-
-  def destroy
-    @user = current_user
-    @activity = Activity.find(params[:id])
-    if Activity.destroy
-      redirect_to activity_path
-    else
-      render :index
-    end
-  end
-
   def edit
     @activity = Activity.find(params[:id])
+    authorize @activity
   end
 
   def update
     @activity = Activity.find(params[:id])
     @activity.update(activity_params)
+    authorize @activity
+  end
+
+  def destroy
+    # @user = current_user
+    @activity = Activity.find(params[:id])
+    authorize @activity
+
+    if Activity.destroy
+      redirect_to activity_path
+    else
+      render :index
+    end
   end
 
   private
@@ -47,6 +59,6 @@ class ActivitiesController < ApplicationController
   end
 
   def activity_params
-    params.require(:activity).permit(:name, :description, :category)
+    params.require(:activity).permit(:name, :description, :category, :date, :address, :limit)
   end
 end
